Refactor NMEA utils into their own class
* GpsTestUtil turned into a catch-all for utilities, so let's organize them a bit better
