+++ GPSTest/src/main/java/com/android/gpstest/GpsStatusFragment.java
@@ -47,6 +47,7 @@ import com.android.gpstest.model.SatelliteStatus;
 import com.android.gpstest.util.CarrierFreqUtils;
 import com.android.gpstest.util.GpsTestUtil;
 import com.android.gpstest.util.MathUtils;
+import com.android.gpstest.util.NmeaUtils;
 import com.android.gpstest.util.PreferenceUtils;
 import com.android.gpstest.util.SortUtil;
 import com.android.gpstest.util.UIUtils;
@@ -465,7 +466,7 @@ public class GpsStatusFragment extends Fragment implements GpsTestListener {
             return;
         }
         if (message.startsWith("$GPGGA") || message.startsWith("$GNGNS") || message.startsWith("$GNGGA")) {
-            Double altitudeMsl = GpsTestUtil.getAltitudeMeanSeaLevel(message);
+            Double altitudeMsl = NmeaUtils.getAltitudeMeanSeaLevel(message);
             if (altitudeMsl != null && mNavigating) {
                 if (mPrefDistanceUnits.equalsIgnoreCase(METERS)) {
                     mAltitudeMslView.setText(mRes.getString(R.string.gps_altitude_msl_value_meters, altitudeMsl));
@@ -475,7 +476,7 @@ public class GpsStatusFragment extends Fragment implements GpsTestListener {
             }
         }
         if (message.startsWith("$GNGSA") || message.startsWith("$GPGSA")) {
-            DilutionOfPrecision dop = GpsTestUtil.getDop(message);
+            DilutionOfPrecision dop = NmeaUtils.getDop(message);
             if (dop != null && mNavigating) {
                 showDopViews();
                 mPdopView.setText(mRes.getString(R.string.pdop_value, dop.getPositionDop()));
