+++ GPSTest/src/test/java/com/android/gpstest/NmeaUtilsTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016 Sean J. Barbeau (sjbarbeau@gmail.com)
+ * Copyright (C) 2013-2019 Sean J. Barbeau
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,15 +15,14 @@
  */
 package com.android.gpstest;
 
-import com.android.gpstest.util.GpsTestUtil;
+import com.android.gpstest.util.NmeaUtils;
 
 import org.junit.Test;
 
 import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertNull;
 
-public class GpsTestUtilTest {
-
+public class NmeaUtilsTest {
     /**
      * Test getting altitude above mean sea level (geoid) from NMEA sentences
      */
@@ -37,21 +36,21 @@ public class GpsTestUtilTest {
         final String gnGga1 = "$GNGGA,114926.00,3206.341435,N,11850.092448,E,1,11,0.9,19.9,M,2.2,M,,*7E";
         final String gnGga2 = "$GNGGA,172814.00,2803.208136,N,08225.981423,W,1,08,1.1,-19.7,M,-24.8,M,,*5F";
 
-        altitude = GpsTestUtil.getAltitudeMeanSeaLevel(gpsSentence);
+        altitude = NmeaUtils.getAltitudeMeanSeaLevel(gpsSentence);
         assertEquals(19.2d, altitude);
 
-        altitude = GpsTestUtil.getAltitudeMeanSeaLevel(gnssSentence);
+        altitude = NmeaUtils.getAltitudeMeanSeaLevel(gnssSentence);
         assertEquals(78.9d, altitude);
 
-        altitude = GpsTestUtil.getAltitudeMeanSeaLevel(gnGga1);
+        altitude = NmeaUtils.getAltitudeMeanSeaLevel(gnGga1);
         assertEquals(19.9d, altitude);
 
-        altitude = GpsTestUtil.getAltitudeMeanSeaLevel(gnGga2);
+        altitude = NmeaUtils.getAltitudeMeanSeaLevel(gnGga2);
         assertEquals(-19.7d, altitude);
 
         final String badGnssSentence
                 = "$GNGNS,015002.0,2804.733672,N,08224.631117,W,AAN,09,1.1,BAD,-24.0,,*23";
-        altitude = GpsTestUtil.getAltitudeMeanSeaLevel(badGnssSentence);
+        altitude = NmeaUtils.getAltitudeMeanSeaLevel(badGnssSentence);
         assertNull(altitude);
     }
 
@@ -76,37 +75,37 @@ public class GpsTestUtilTest {
         // From http://www.gpsinformation.org/dale/nmea.htm#GSA
         final String s6 = "$GPGSA,A,3,04,05,,09,12,,,24,,,,,2.5,1.3,2.1*39";
 
-        dop = GpsTestUtil.getDop(s1);
+        dop = NmeaUtils.getDop(s1);
 
         assertEquals(1.3d, dop.getPositionDop());
         assertEquals(1.0d, dop.getHorizontalDop());
         assertEquals(0.8d, dop.getVerticalDop());
 
-        dop = GpsTestUtil.getDop(s2);
+        dop = NmeaUtils.getDop(s2);
 
         assertEquals(3.6d, dop.getPositionDop());
         assertEquals(1.8d, dop.getHorizontalDop());
         assertEquals(3.1d, dop.getVerticalDop());
 
-        dop = GpsTestUtil.getDop(s3);
+        dop = NmeaUtils.getDop(s3);
 
         assertEquals(3.6d, dop.getPositionDop());
         assertEquals(1.8d, dop.getHorizontalDop());
         assertEquals(3.1d, dop.getVerticalDop());
 
-        dop = GpsTestUtil.getDop(s4);
+        dop = NmeaUtils.getDop(s4);
 
         assertEquals(3.6d, dop.getPositionDop());
         assertEquals(2.1d, dop.getHorizontalDop());
         assertEquals(2.2d, dop.getVerticalDop());
 
-        dop = GpsTestUtil.getDop(s5);
+        dop = NmeaUtils.getDop(s5);
 
         assertEquals(1.7d, dop.getPositionDop());
         assertEquals(1.0d, dop.getHorizontalDop());
         assertEquals(1.3d, dop.getVerticalDop());
 
-        dop = GpsTestUtil.getDop(s6);
+        dop = NmeaUtils.getDop(s6);
 
         assertEquals(2.5d, dop.getPositionDop());
         assertEquals(1.3d, dop.getHorizontalDop());
